requirements.txt
# --- Core API ---
fastapi
uvicorn
pydantic
python-dotenv
anthropic==0.25.0
pydantic-settings==2.13.1
python-multipart==0.0.9
httpx==0.27.0
aiofiles==23.2.1
slowapi==0.1.9
loguru==0.7.2

# --- LangChain ---
langchain-groq
langchain-community
langchain-core
langchain-google-genai
tavily-python

# --- Database ---
duckdb
duckdb-engine==0.15.0
sqlalchemy==2.0.36
pandas

# --- Vector Stores ---
qdrant-client==1.17.1
chromadb==1.5.5
protobuf>=5,<8
sentence-transformers

# --- Document parsing ---
pypdf

# --- Plotting (used by plotting_agent at runtime) ---
matplotlib
seaborn
numpy

# --- ML / Analytics ---
scikit-learn==1.7.2
